Use PSR Http Client instead of Guzzle (#50)
* add php-http/message library to get psr7 interfaces

* replace guzzle to psr7 http client

* add request factory to test ec2 detector

* replace guzzle to psr7 in eks detector

* add request factory to eks detector test

* Supress UndefinedDocblockClass in ContainerBuilder

Refference to issue: https://github.com/vimeo/psalm/issues/7738
Symfony isn't waiting to implement the new PHP features
until they limit they range versions to fully support those
features. The PHP Enums are not compatible until PHP 8.1
so the Psalm is reporting the issue, that UnitEnum type not
exist. There are no fix for this issue yet in Psalm or in
Psalm Symfony Plugin. Only way to fix it at this point is
to suppress the error but limit that to specific class and
specific file in code. This is done in this commit.