Refactor to use phpstan symfony plugin
There is a problem in phpstan with getting the right type of root node
from tree builder. It can be fixed using phpstan symfony plugin, but it
requires to instantiate TreeBuilder in place where you get root node, to
allow phpstan to check second parameter of TreeBuilder. This change is
fixing some problems that was ignored before.