Refactor/static analyse improvement (#51)
* Update phpstan and install phpstan symfony plugin

* Add phpstan extension to configuration

* Remove unused property from test class

* Change static:: to self:: for const refs

* Remove unecessery psalm suppressions

* Refactor to use phpstan symfony plugin

There is a problem in phpstan with getting the right type of root node
from tree builder. It can be fixed using phpstan symfony plugin, but it
requires to instantiate TreeBuilder in place where you get root node, to
allow phpstan to check second parameter of TreeBuilder. This change is
fixing some problems that was ignored before.

* Replace TreeBuilder instancing to shorter notation